fix for different profile in nested directory
The zsh for loop loops through an associative array in alphabetical
order. It prevents to use different dirprofile in a directory which one
of it's parent has dir profile too, because the less specific will come
first.

Instead of using associative array, use regular array to define
directory profiles. That keeps the order, so if an item is placed before
another, that will be called.